Move plotdebug to utility
